#OSIS-2956 Nettoyage redondance cyclique
